All Altitudes set to deterministic values
I've pretty much had it with Way-point Generator and altitude settings. From now on all flight altitude waypoints with the exception of TSA15 will be deterministic and set to 0.

This will make the Island Airport seem like the terminal is underground, however it is well within the 50 meter distance to complete flight missions.

All TSA15 flights have been gathered under KSIS contracts and the dst altitude is set to 133 for the TSA15 site and 0 for all others.

I'll entertain no more requests to try and sort this out,