Update 09 - AllBases.cfg.txt
This automatic contract file can be used to open all airports without having to do all of the landing rights flights.

Rename this to "09 - AllBases.cfg"  It will execute automatically after Let's Start an Airline and complete Flight Progress for all airports.  You can also enable it any time during to jump to the end of the Landing Rights flights.